Fix Next.js page export validation
Move preserved writer implementation to WriterPage.tsx so page.tsx only exports the redirect, complying with Next.js page file requirements.
